Separate joyride steps and solve linting
